// src/types/weather.d.ts
export type Weather =
  | 'Thunderstorm'
  | 'Drizzle'
  | 'Rain'
  | 'Snow'
  | 'Mist'
  | 'Smoke'
  | 'Haze'
  | 'Dust'
  | 'Fog'
  | 'Sand'
  | 'Ash'
  | 'Squall'
  | 'Tornado'
  | 'Clear'
  | 'Clouds';

/*
Thunderstorm 雷雨
Drizzle 霧雨
Rain 雨
Snow 雪
Mist 霧
Smoke 煙
Haze 薄霧
Dust 砂煙
Fog 霧
Sand 砂嵐？ 砂という意味から意訳して砂嵐だと思った
Ash 火山灰
Squall スコール
Tornado 竜巻
Clear 晴れ
Clouds 曇り

from: https://openweathermap.org/weather-conditions
*/